hack torch test into learn.py
