+++ .yamato/training-int-tests.yml
@@ -14,11 +14,6 @@ test_mac_training_int_{{ editor.version }}:
   commands:
     - pip install pyyaml
     - python -u -m ml-agents.tests.yamato.training_int_tests
-    # Backwards-compatibility tests.
-    # If we make a breaking change to the communication protocol, these will need
-    # to be disabled until the next release.
-    - python -u -m ml-agents.tests.yamato.training_int_tests --python=0.15.0
-    - python -u -m ml-agents.tests.yamato.training_int_tests --csharp=0.15.0
   dependencies:
     - .yamato/standalone-build-test.yml#test_mac_standalone_{{ editor.version }}
   triggers:
