+++ ml-agents/setup.py
@@ -60,7 +60,7 @@ setup(
         "numpy>=1.13.3,<2.0",
         "Pillow>=4.2.1",
         "protobuf>=3.6",
-        "pytorch>=1.0.2",
+        "torch>=1.0.2",
         "pyyaml",
         "tensorflow>=1.7,<2.1",
         'pypiwin32==223;platform_system=="Windows"',
